Cleaned up types/message_types #85
