+++ cpp/include/krypto/utils/common.h
@@ -7,10 +7,68 @@
 #include <cmath>
 #include <spdlog/fmt/fmt.h>
 #include <ostream>
-#include <krypto/utils/types.h>
 #include <krypto/serialization/serialization_generated.h>
 
 namespace krypto::utils {
+
+    enum class MsgType : uint8_t {
+        UNDEFINED,
+        NO_PAYLOAD,
+        QUOTE,
+        TRADE,
+        HEARTBEAT,
+        INSTRUMENT_REQUEST,
+        INSTRUMENT_RESPONSE,
+        INSTRUMENT_CACHE_REFRESH_REQUEST,
+        ORDER,
+        ORDER_REQUEST,
+        ORDER_CANCEL_REQUEST,
+        ORDER_REPLACE_REQUEST,
+        ORDER_UPDATE,
+        POSITION,
+        RISK_SUMMARY,
+        THEO,
+        SIZE
+    };
+
+    static constexpr const char *MsgTypeNames[enum_size<MsgType>()] = {
+            "",
+            "AA",
+            "MQ",
+            "MT",
+            "HH",
+            "IQ",
+            "IS",
+            "IC",
+            "O_",
+            "OO",
+            "OC",
+            "OR",
+            "OU",
+            "PP",
+            "PR",
+            "PT"
+    };
+
+    static constexpr const MsgType MsgTypes[enum_size<MsgType>()] = {
+            MsgType::UNDEFINED,
+            MsgType::NO_PAYLOAD,
+            MsgType::QUOTE,
+            MsgType::TRADE,
+            MsgType::HEARTBEAT,
+            MsgType::INSTRUMENT_REQUEST,
+            MsgType::INSTRUMENT_RESPONSE,
+            MsgType::INSTRUMENT_CACHE_REFRESH_REQUEST,
+            MsgType::ORDER,
+            MsgType::ORDER_REQUEST,
+            MsgType::ORDER_CANCEL_REQUEST,
+            MsgType::ORDER_REPLACE_REQUEST,
+            MsgType::ORDER_UPDATE,
+            MsgType::POSITION,
+            MsgType::RISK_SUMMARY,
+            MsgType::THEO
+    };
+
     constexpr const int64_t exchange_base = 100;
 
     constexpr auto exchange_id(const krypto::serialization::Exchange et) {
@@ -33,5 +91,5 @@ namespace krypto::utils {
 
     std::unordered_map<std::string, krypto::serialization::Currency> name_to_currency() noexcept;
 
-    std::unordered_map<std::string, krypto::utils::MsgType> name_to_msg_type();
+    std::unordered_map<std::string, MsgType> name_to_msg_type();
 }
