+++ cpp/include/krypto/utils/types.h
@@ -7,6 +7,7 @@
 #include <ostream>
 #include <krypto/utils/enum.h>
 #include <krypto/serialization/serialization_generated.h>
+#include <krypto/utils/common.h>
 
 
 namespace krypto::utils {
@@ -15,64 +16,6 @@ namespace krypto::utils {
     using secid_t = int64_t;
     using timestamp_t = int64_t;
     using seqnum_t = int64_t;
-
-    enum class MsgType : uint8_t {
-        UNDEFINED,
-        NO_PAYLOAD,
-        QUOTE,
-        TRADE,
-        HEARTBEAT,
-        INSTRUMENT_REQUEST,
-        INSTRUMENT_RESPONSE,
-        INSTRUMENT_CACHE_REFRESH_REQUEST,
-        ORDER,
-        ORDER_REQUEST,
-        ORDER_CANCEL_REQUEST,
-        ORDER_REPLACE_REQUEST,
-        ORDER_UPDATE,
-        POSITION,
-        RISK_SUMMARY,
-        THEO,
-        SIZE
-    };
-
-    static constexpr const char *MsgTypeNames[enum_size<MsgType>()] = {
-            "",
-            "AA",
-            "MQ",
-            "MT",
-            "HH",
-            "IQ",
-            "IS",
-            "IC",
-            "O_",
-            "OO",
-            "OC",
-            "OR",
-            "OU",
-            "PP",
-            "PR",
-            "PT"
-    };
-
-    static constexpr const MsgType MsgTypes[enum_size<MsgType>()] = {
-            MsgType::UNDEFINED,
-            MsgType::NO_PAYLOAD,
-            MsgType::QUOTE,
-            MsgType::TRADE,
-            MsgType::HEARTBEAT,
-            MsgType::INSTRUMENT_REQUEST,
-            MsgType::INSTRUMENT_RESPONSE,
-            MsgType::INSTRUMENT_CACHE_REFRESH_REQUEST,
-            MsgType::ORDER,
-            MsgType::ORDER_REQUEST,
-            MsgType::ORDER_CANCEL_REQUEST,
-            MsgType::ORDER_REPLACE_REQUEST,
-            MsgType::ORDER_UPDATE,
-            MsgType::POSITION,
-            MsgType::RISK_SUMMARY,
-            MsgType::THEO
-    };
 
     struct Heartbeat {
         timestamp_t timestamp;
@@ -137,4 +80,66 @@ namespace krypto::utils {
         std::map<int64_t, int64_t> asks;
         Quote quote;
     };
+
+    struct OrderUpdate final {
+        static const constexpr krypto::utils::MsgType message_type =
+                krypto::utils::MsgType::ORDER_UPDATE;
+        int64_t timestamp;
+        std::string order_id;
+        std::string exchange_order_id;
+        krypto::serialization::OrderStatus status;
+        int64_t filled_quantity;
+    };
+    struct OrderRequest final {
+        static const constexpr krypto::utils::MsgType message_type =
+                krypto::utils::MsgType::ORDER_REQUEST;
+        int64_t timestamp;
+        int64_t security_id;
+        int64_t price;
+        int64_t quantity;
+        krypto::serialization::Side side;
+        std::string order_id;
+        krypto::serialization::TimeInForce tif;
+    };
+    struct OrderCancelRequest final {
+        static const constexpr krypto::utils::MsgType message_type =
+                krypto::utils::MsgType::ORDER_CANCEL_REQUEST;
+        int64_t timestamp;
+        std::string order_id;
+    };
+    struct OrderReplaceRequest final {
+        static const constexpr krypto::utils::MsgType message_type =
+                krypto::utils::MsgType::ORDER_REPLACE_REQUEST;
+        int64_t timestamp;
+        std::string order_id;
+        int64_t price;
+        int64_t quantity;
+        krypto::serialization::Side side;
+    };
+
+    struct TheoreticalSnapshot final {
+        static const constexpr krypto::utils::MsgType message_type =
+                krypto::utils::MsgType::THEO;
+        int64_t timestamp;
+        int64_t security_id;
+        double_t price;
+        double_t adjusted_price;
+        double_t error;
+        double_t mm_base_bid;
+        double_t mm_base_ask;
+        double_t mm_base_spread;
+        double_t bid_liquidity;
+        double_t ask_liquidity;
+        bool valid;
+
+        friend std::ostream &operator<<(std::ostream &os, const TheoreticalSnapshot &snapshot) {
+            os << "security_id: " << snapshot.security_id << " price: " << snapshot.price << " adjusted_price: "
+               << snapshot.adjusted_price << " error: " << snapshot.error << " mm_base_bid: " << snapshot.mm_base_bid
+               << " mm_base_ask: " << snapshot.mm_base_ask << " mm_base_spread: " << snapshot.mm_base_spread
+               << " bid_liquidity: " << snapshot.bid_liquidity << " ask_liquidity: " << snapshot.ask_liquidity
+               << " valid: " << snapshot.valid;
+            return os;
+        }
+    };
+
 }
